pyproject.toml
[tool.poetry]
name = "user-management-api"
version = "0.1.0"
description = "REST API for user management with Litestar and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "app", from = "." }
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
pydantic = "2.11.4"
litestar = {extras = ["standard"], version = "^2.16.0"}
litestar-granian = "^0"
litestar-asyncpg = "^0"
advanced-alchemy = "1.4.2"
msgspec = "^0.19.0"
psycopg2-binary = "^2.9"
sqladmin-litestar-plugin = "^0.2.0"
passlib = "^1.7.4"
pydantic-settings = "^2.9.1"
loguru = "^0.7.3"
alembic = "^1.15.2"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"